+++ examples/expansions/try_join_cyclic.rs
@@ -206,6 +206,7 @@ fn expansion() {
     let _fut_3 = match (async { Some(()) }, async { Some(()) }, async { Some(()) }) {
         futs => {
             use ::core::future::Future;
+            use ::core::hint::unreachable_unchecked;
             use ::core::ops::ControlFlow;
             use ::core::option::Option::{self, None, Some};
             use ::core::pin::Pin;
@@ -214,11 +215,6 @@ fn expansion() {
             use ::core::task::{Context, Poll};
 
             try_trait!();
-
-            #[inline(always)]
-            unsafe fn unreachable_unchecked() -> ! {
-                ::core::unreachable!("`unreachable_unchecked` reached at runtime")
-            }
 
             enum MaybeDone<F: Future>
             where
@@ -314,43 +310,32 @@ fn expansion() {
 
                     if !unsafe {
                         const COUNT: ::core::primitive::u8 = 3;
-                        let mut done = true;
                         let to_skip =
                             ::core::mem::replace(skip_next_time, (*skip_next_time + 1) % COUNT);
 
                         use ::core::iter::Iterator;
-                        if let ControlFlow::Break(r) = Iterator::try_for_each(
+                        match Iterator::try_fold(
                             &mut Iterator::chain(to_skip..COUNT, 0..to_skip),
-                            |i| {
-                                match i {
-                                    0 => {
-                                        done &= MaybeDone::poll(
-                                            Pin::new_unchecked(&mut *maybe_done0),
-                                            cx,
-                                        )?
-                                    }
-                                    1 => {
-                                        done &= MaybeDone::poll(
-                                            Pin::new_unchecked(&mut *maybe_done1),
-                                            cx,
-                                        )?
-                                    }
-                                    2 => {
-                                        done &= MaybeDone::poll(
-                                            Pin::new_unchecked(&mut *maybe_done2),
-                                            cx,
-                                        )?
-                                    }
-                                    _ => unsafe { unreachable_unchecked() },
-                                }
-
-                                ControlFlow::Continue(())
+                            true,
+                            |done, i| match i {
+                                0 => ControlFlow::Continue(
+                                    MaybeDone::poll(Pin::new_unchecked(&mut *maybe_done0), cx)?
+                                        && done,
+                                ),
+                                1 => ControlFlow::Continue(
+                                    MaybeDone::poll(Pin::new_unchecked(&mut *maybe_done1), cx)?
+                                        && done,
+                                ),
+                                2 => ControlFlow::Continue(
+                                    MaybeDone::poll(Pin::new_unchecked(&mut *maybe_done2), cx)?
+                                        && done,
+                                ),
+                                _ => unsafe { unreachable_unchecked() },
                             },
                         ) {
-                            return Poll::Ready(Try::from_residual(r));
+                            ControlFlow::Break(r) => return Poll::Ready(Try::from_residual(r)),
+                            ControlFlow::Continue(done) => done,
                         }
-
-                        done
                     } {
                         return Poll::Pending;
                     }